Latest Version
I have missed one bug in previos version now its ok, but when you check it please tell me about any glitches.
AND 1 important notice: there is an error when you drag the mouse to the black space. This space is not a part of the level and appears because of isometric view, but as I haven't done the appropriate camera clamping it is visible. So just try avoid dragging mouse to it. I will fix it later or I make the "Bounds bigger"